requirements.txt
langchain>=0.3
langchain-anthropic>=0.3
langchain-core>=0.3
anthropic>=0.40
sqlalchemy>=2.0
geoalchemy2>=0.15
psycopg2-binary>=2.9
python-dotenv>=1.0
streamlit>=1.40
rich>=13.0
mcp[cli]>=1.0
